Fix file reconstruction if the first object is invalid
The first object in a file is the only one that HexaPDF will try to read
to determine whether the file is linearized. If that first object is
invalid, the tokenizer position after trying to read the invalid object
may be anything. This means that when determining the end of the first
object we might wrongly have skipped good, readable objects.

Fix this by making saving the cursor position before reading the first
object and then restoring it before trying to find the end of it.