Fix GraphicObject::EndpointArc to correctly determine the start/end points
The current implementation follows the spec layed out by SVG. However,
when doing round-trip tests, i.e. creating an arc and from its values
and endpoint arc which in turn creates an arc internally, the start and
end points do not match.

The reason for this is a division of x by r_x and y by r_y of two
vectors (that determine the start and end angles) which, when r_x and
r_y are different, change the orientation of the vectors and thus the
angles.

We can fix this by *not* doing those divisions although then our
implementation doesn't follow the SVG spec anymore.