Allow comparison of HexaPDF::Object to simple value if not indirect
If a PDF object is not indirect, it just wraps a simple value like an
integer or array. However, comparison with #== doesn't show that and
will fail, ie. `HexaPDF::Object.new(1) != 1`.

The Rectangle class already has a special implementation of #== which
allows such a comparison. This change implements that for all objects.

This is a major change in one of the most used classes of HexaPDF.
However, there should not be any negative impact due to this change.