Add page resource pruning to the optimization task
The new option allows one to remove all XObjects from a page's resources
dictionary that are not referenced from any page content stream.

This is needed because some PDF writers use a single resources
dictionary for all pages. So when a single page is extracted into a new
PDF the resources of all pages will also be found in the new PDF.