Add support for signing a PDF with the adbe.pkcs7.detached algorithm
Reading and validating a signature is already supported, the missing
part is the creation of a digital signature.

With this change the foundation for signing PDFs is established. Only
one algorithm, adbe.pkcs7.detached, is currently supported. However,
using that signing handler for ETSI.CAdES.detached should also work.

By extracting the specific parts of signing into a handler, it should be
easily possible to customize signing. E.g. using an external signing
service for generating the needed PKCS7 structure.

This is only a barebones implementation, doing the minimum necessary
things to create a valid digital signature. More convenience methods as
well as support for PDF 2.0 specific parts (like LTV signatures) will
probably need to be added.