Fix scaled font size calculation when using Type3 fonts
The scaled font size calculation in GraphicsState and Layout::Style is
correct for all fonts but Type3 fonts. The reason for this is that Type3
fonts can have an aribtrary matrix for transforming glyph space to text
space whereas all other fonts just use a scaling matrix of the form
[0.001 0 0 0.001 0 0] (i.e. scaling by dividing through 1000).

Adapt the calculation to also be correct when Type3 fonts are used.
Correct in this case, however, means that we simplify the font matrix of
Type3 fonts and only think of it as a simple scaling matrix. This should
be fine in (nearly) all real-world cases.