Use array with two empty strings for trailer ID field in case it is missing when decrypting
The specification says that the first string of the trailer's ID field
needs to be part of the computation of the user encryption key. It
doesn't, however, mention that the ID field is mandatory which was kinda
assumed.

Since some PDF writers don't add an ID field and do the encryption
calculations without it, we need to use an empty string in such cases.