Fix incremental writing in case of existing cross-reference streams
If a PDF file contains cross-reference streams and a new revision is
appended, it also needs to have a cross-reference stream and not a
cross-reference table.

This doesn't seem to be a problem for most open source PDF viewers and
applications but some proprietary applications will mark the file as
damaged.

So we make sure that if there is at least one existing cross-reference
stream that all following revisions also use cross-reference streams.