Associate page with field widget when creating it
An annotation must be associated with a page by listing it in the
/Annots key of the page.

Additionally, there is the /P key of the annotation itself that allows
associating the page with the annotation. This key is not mandatory in
nearly all cases. However, it seems that Acrobat Reader somehow needs it
to correctly display the signature panel, at least if it is a visible
signature...