Fix incremental writing in case of deleted object streams
If an object stream is deleted (e.g. when optimizing the file), the oid
is marked as free in the added revision. This makes the original objects
stored in the object stream unavailable and thus the file invalid.

Fix this by ignoring deleted object and cross-reference streams when
detecting modified objects. This should do no harm since both object
types are not used for content but for the PDF structure itself.