Fix Revision#delete to re-use the deleted data object
When currently deleting an object using `mark_as_free: true`, a
completely new object is inserted in place of the deleted one in the
revision. So when comparing the deleted object to the object retrieved
via Revision#object, the result is false.

The result should be true because they are in fact the same object. By
re-using the Object#data object, this is fixed.