Fix Type::Annotation#appearance to correctly return wrapped object
Some invalid files contain appearance streams that are missing the /Type
and /Subtype keys. While /Type is optional, /Subtype is required.

Due to the missing keys, the object is just wrapped in the standard
Stream class instead of the correct Type::Form class.

This situation alone is already correctly handled. However, if the
appearance is part of a text field and the appearance is regenerated,
the appearance object is re-used and the missing values are set. So now
we have a Stream object that, if wrapped again, would be wrapped in a
Form class.

However, the code in #appearance just checks for the correct values for
/Type and /Subtype, and since these are now set, the Stream object is
returned (leading to errors later on).

We need to additionally check that the object is not an instance of the
Stream class to prevent this.