Fix decrypting of files with multiple revisions
If a PDF is encrypted and has multiple revisions, it also contains
multiple /Encrypt dictionaries. All these dictionaries contain the same
data which is not encrypted.

The current code only treats the /Encrypt dictionary of the latest
version specially, but not the others, leading to decryption errors.

Fix this by treating all encryption dictionaries the same.