Fix regression due to change in HexaPDF::Object#== when using Type1 fonts
When a Type1 font like Helvetica is added to a document, a PDF object is
created but not immediately added to the document. Due to the recent
change in HexaPDF::Object#==, different fonts, e.g. one standard and one
with a custom encoding, were deemed one and the same.

Fix this by using Document#add instead of #wrap. This would later be
done anyways, so it is no problem.