Refactor Composer (style support, API docs)
The current way of using styles with HexaPDF::Composer is quite
cumbersome since most of the steps have to be done by the user.

This commit introduces a new method for defining and storing styles,
Composer#style, that replace the #base_style method. This way it is
possible to predefine all needed styles directly on the Composer object
and to reference those styles by name.

Furthermore, usage of styles by the various methods (e.g. #text) has
been made more consistent.