Remove caching of document config option in font wrappers
The caching of the configuration option font.on_missing_glyph was
initially added to avoid storing a reference to the document in font
wrappers.

However, since a font wrapper always has a reference to the document
through the pdf_object variable, this caching is not necessary. In fact,
it makes problems if a font is first added to the document and then only
afterwards the configuration option changed.