Refactor Document and Revisions with respect to object handling
The Revisions class is currently responsible for managing the revisions
of a document and the Document class for managing the objects.

However, the current interfaces make it too easy to create an invalid
state. For example, when adding an object to another than the current
revision.

This change moves the object handling into the Revisions class so as to
make sure that creating something invalid is not easily possible. The
Document interface now mostly defers to Revisions when handling objects.

Additionally changes for other issues (like invalidly created files due
to manually adding a revision and then modifying an existing object)
will be addressed in future commits.