Fix Writer to move modified objects into the current revision
When writing a file with two or more revisions, it is necessary to write
modified objects always into the current revision.

For example, when loading a PDF file, adding a revision, and then
inserting a page, the page object will be in the new revision. However,
the modified pages object will be in another revision, referring to an
invalid object number.