Fix handling of hybrid-reference files
For a hybrid-reference file the current implementation merges the data
from the XRefStm cross-reference stream with the data from the previous
revision.

However, the XRefStm cross-reference stream is actually part of the
current revision. So we need to merge it with the current revision and
make sure that the cross-reference table takes precedent.