Set /Type to /XRef for cross-reference stream trailers
When extracting the trailer information from an xref stream, the /Type
key is not set. So there is no possibility to know whether that trailer
came from an xref table or stream.

By setting the /Type key we can provide this information. The #type
method for the trailer will still return :XXTrailer as expected.