Change Layout::TextBox to use available height when valigning to center/bottom
When aligning some text to the center or the bottom of the available
region, the resulting box should use the whole available height as box
height.

This works for bottom alignment due to the y-offsets set by the
TextLayouter but not for center alignment. There the box will end just
after the text, so seems to be bottom aligned.

To make sure that this always works regardless of the inner workings of
TextLayouter we set the box height to the available height when aligning
to center or bottom.