Add Layout::MultiFrame for layouting multiple boxes in multiple frames
We often need to place boxes into two or more frames (e.g. with column
layout). The algorithm for doing so is the same regardless of the
concrete application.

The MultiFrame class implements this behaviour and it is expected that
it will be used in the composer as well as in container boxes.