Fix Layout::TextBox#split to work for position :flow in special case
When using position :flow for a text box, the situation may arise that
the initial available width is smaller than some width below it. E.g.

        o----
        |   |
    -----   |
    |       |
    ---------

The box will then be assigned the larger width which is larger than the
initial available width.

Due to this such text boxes weren't split correctly.

This commit fixes this by explicitly checking for boxes with position
:flow where the failure reason was not enough height and split them
correctly.