Change Layout::Frame#trim_shape to use a better algorithm
The current way #trim_shape does it's job is very straight-forward but
doesn't provide good results in some cases.

One problem is shapes of the following form:

    o----y--------
    |            |
    |            |
    -----x       |
         |       |
         ---------

In such a case the first region's top-left corner would be at o and the
second region's one would be at x. However, position y is much better
because otherwise there would be an empty horizontal slice.

This commit changes the algorithm of #trim_shape so that the second
region will start at y instead of x.