Extract box creation functionality from Composer into own class
The #text, #formatted_text and #image methods of Composer create the
boxes and also draw them. However, it is useful to have the box creation
functionality at a common place (for example, when the composer is not
used.)

This commit extracts the guts of those methods and puts them into a new
convenience class, available through 'document.layout'.