Update box implementations to indicate whether they support position :flow
If a box doesn't support the value :flow of the style property
'position' but it is set nonetheless, the result is wrong (because Frame
thinks it works but the box acts otherwise).

This problem won't occur if one is strict with setting the position
property.

However, to make this easier and more in line with what one would
expect, the method #supports_position_flow? is introduced which
indicates whether the box supports flowing its content.