Fix Document::Signatures#add to work when signature object is last object written
Currently, the Writer returns just the xref section when writing
incrementally. This allows one to get the offset and length of every
object *except* the last one; there the length is missing because we
have no offset of the content after the object.

However, we *do* actually have that offset already, it is the offset of
the start of the cross-reference section. By returning that offset and
incorporating it into the list of offsets, the method works even if the
signature object is the last object written.