Add Document::Layout::ChildrenCollector for collecting children boxes
Container boxes need their children to be specified before they can be
fitted and drawn. Currently, one needs to separately specify those child
boxes and then pass the :children keyword argument.

To make this more seamless, the ChildrenCollector class is introduced.
Its methods allow specifying boxes that are collected internally and can
be fetched using the #children method.

An instance of this class is yielded for the Layout#box method if a
block is given. This way one can specify a container box and the
children boxes in the provided block.