Fix incremental writing to only use xref streams if a revision used one
At least one major proprietary application fails to validate a signed
document that consists of multiple revisions where all revisions are
written using cross-reference tables except the last one with the
signature which is written with a cross-reference stream. A document
structured like this works fine in all tested open source viewers.

A similar problem was already addressed in a prior commit where it was
the other way around (all except the last one used cross-reference
streams).

The current code just tests for a loaded cross-reference stream. However,
if a revision contained an `/XRefStm` entry, a cross-reference stream
would be loaded, although not as the main reference for a revision, and
thus a cross-reference stream written.

This is now fixed so that a cross-reference stream is only written if a
revision actually used one, not just if an `/XRefStm` entry is present.