Fix reading of glyph list on Windows
Windows has this bad habit of distinguishing between text and binary
files. If a binary file is accidentally read in text mode, you will get
a real mess.

The glyph list file is not really a binary file but one with UTF-8
characters. And some of the characters are from the control plane and
those seem to be messed up on Windows. And messed up so bad that most of
the part of the file is just not read.

Using the binary mode and tagging the read strings with UTF-8 fixes
this.