Fix parsing error that occurs due to an invalid stream length
Currently, an error is raised when parsing a file that contains a stream
whose length value is invalid and where the supposed end is not followed
by the keyword 'endstream' but by an invalid PDF object.

We can infer from the invalid PDF object that the length value is not
valid and thus initiate the stream recovery algorithm.