Fix HexaPDF::Object#== comparison of non-indirect objects
Currently, if a non-indirect object is compared to another Object,
another call to #== is made on the other Object. If the other object is
also a non-indirect object, the effect is that the values of the two are
compared. So two different HexaPDF::Object instances may compare as the
same if both are indirect and have the same value.

This might be a problem because during writing an object number may be
assigned and then those two objects would not be the same anymore!

So we need to restrict the comparison of non-indirect objects to other
objects that are not HexaPDF::Object instances. And thus only define two
HexaPDF::Object instances as the same if they contain the same #data
object.