Fix Type::PageTreeNode#perform_validation to only work on the document's root node
The page tree node validation already only works for root nodes, i.e.
ones without a /Parent entry. However, this may lead to problems if
there are multiple such nodes (e.g. due to restructuring the page tree
or due to incremental updates) because if a root node that is not the
document's root node and which has references to current pages gets
validated later than the actual document root node, those pages may get
modified to contain invalid state.

So now we only do the extended validation if the page tree node is the
document's root node.