Fix Type::Page to only perform validation when in the document's page tree
A page's validation is currently done unconditionally. This is a problem
if the page is one that got removed from the document's page tree but
still available. And for artifact property lists <</Type /Page>> which
appear to be a page node but are actually not.

To fix this we now only validate pages that are part of the document's
page tree.