Fix SecurityHandler#encrypt_string to return a dupped string
The method normally returns a string different from the input, namely
the encrypted version of it. However, sometimes no encryption must be
performed in such a case we now make sure to return a dupped string.

This allows callers, especially Serializer#serialize_string, to treat
the result as a string that can be modified without repercussions.