Fix Layout::TextLayouter#fit to correctly handle non-zero width breakpoint penalties
If a non-zero width breakpoint penalty like a soft-hyphen is encountered
where the box before it fits but the penalty item doesn't and it has a
box after it, the fitting algorithm hangs in an infinite loop because
the box before already fits and so the height is not changed to find a
new location.

This is fixed by explicitly checking this situation and making it a
box-too-wide one.