Fix Type::ObjectStream to parse initial stream data right after initialization
Under certain circumstances it is possible that #write_objects is called
before #parse_stream. When that happens, the /N and /First entries are
changed and invalid when #parse_stream is later called, leading to
errors.

This is fixed by making sure that the *initial* stream data is parsed
right after initialization and that it is only parsed once.

A circumstance when this error happens:

1. A PDF with two revisions where the second revision has an empty
   cross-reference table but a /XRefStm entry pointing to a
   cross-reference stream in the initial revision. The object stream is
   also in the initial revision. Both are *not* used by the initial
   revision.

2. The document is just loaded and then written without validation.

This leads to the Writer getting all object streams for the initial
revision and then immediately calling #write_objects. This resets /N and
/First. When the stored objects are accessed, the error happens.