Merge empty revision with /XRefStm into previous one
Sometimes PDF files contain two revisions where the second one contains
a completely empty cross-reference table together with a /XRefStm entry
pointing into the previous revision. So the cross-reference stream is
actually in the first revision but referenced from the second one.

The spec says that the initial revision must not contain a /XRefStm
entry, so it is clear why it is in the second one. However, the second
revision is not really an incremental update because it contains no PDF
objects.

Therefore we merge the second revision into the first one to avoid
errors relating to objects being referenced that are actually in another
revision. This is problematic, for example, when object streams are also
used.