Change Rectangle to recover from invalid values
The Rectangle class is changed so that it is better aligned with how the
parser handles invalid PDF constructs. I.e. if the rectangle doesn't
contain four numbers, parser.on_correctable_error is called and
afterwards the error raised if necessary.

If no error should be raised, the value of the rectangle is reset to [0,
0, 0, 0]. This should do the least damage.