Fix Importer to correctly expose previously mapped objects
When an object has already been imported and is later seen inside
another object, the weak ref is returned instead of the wrapped object.
This leads to serialization issues later.

This is fixed by correctly returning the wrapped object of the weak ref,
like it is already done in other places of the Importer.