Make sure custom properties on TextFragment are taken into account correctly
The TextFragment class is a special in the regard that splitting it in
two, items-wise, and then re-combining it should be possible. Until the
introduction of custom properties, these operations could solely rely on
the assigned style (assigning it to the split fragments or for checking
whether re-combining is possible).

Now the custom properties also need to be taken into account. To make
this easier, some methods are added to TextFragment to facilitate these
two operations.