Fix Revision#each_modified_object to not yield signature objects
Due to a recent change in how HexaPDF::Object#== works, direct objects
are not compared value wise but object identity wise. That makes usually
sense but in the case of #each_modified_object we need value comparison.

This problem has an impact when multiple signature are written since all
signature dictionaries are thought to be modified and therefore yielded.
This leads to invalid files.

Fix this in two ways:

* Make sure that signature dictionaries are never yielded
* Make sure that direct object are compared value-wise