Update signature handler #sign method to require IO+byte range instead of data
Currently, the data to be signed is the only argument of the #sign
method. This could be a problem when signing huge PDF files because then
the contents of the whole PDF is passed to the methods.

Instead, allow the handler to decide whether it needs to load all data
at once by providing it the IO object and the byte range array.