Fix Importer#import to return object of same class as argument
When importing objects, one usually expects that the imported object is
of the same class as the given object. This is true if an object is not
already mapped.

However, if an object was referenced by another, already imported
object, it might have been imported with another class (think:
dictionaries which don't have a /Type entry but are only auto-converted
manually or via dictionary field access).

This change makes sure that the imported object will have the same class
as the argument, regardless of internals.