Simplify Dictionary#perform_validation
The #perform_validation method was changed several years ago and moved
the nested validation of objects to HexaPDF::Object. Due to this change
only required fields or fields with values are now validated.

The old iteration structure, however, was left in place and does more
than needed. This change simplifies the iteration by removing the
iteration method and moving the necessary code directly into
\#perform_validation.