Fix AcroForm::AppearanceGenerator for check boxes to handle nil values
Some PDFs found in the wild have a missing /V value but /AP and /AS
correctly set up, i.e. /AS set to /Off. In such cases the appearance
generator would set /AS to nil which is not correct and may lead to
invalid rendering.

Therefore /AS will be set to /Off in the case of a /V value of nil. This
is also what the appearance generator already does for radio button
fields.