Fix writing of encrypted files containing two or more revisions
When HexaPDF is writing a document, it iterates all revisions, collects
all modified objects and puts them into a new revision. Afterwards it
deletes all active objects from those revisions and then merges the new
revision with the current one.

This leads to the problem for encrypted files with two or more revisions
and an indirect reference to the encryption dictionary that the
encryption dictionary that is stored in doc.trailer[:Encrypt], if not
modified, is deleted from the current revision.

Later when the file is written, the check whether a string belongs to
the encryption dictionary or not fails due to the stored one being
different from the loaded one.

The problem can be fixed by deleting only the modified objecs and not
all active objects.