Create dummy implementation of push button appearance generation
Currently the button field itself raises an error when an appearance for
a push button field should be generated. This doesn't allow the user to
specify custom behaviour as is the case with the field types.

So we move the raising of the error into the appearance generator. Then
a user can define a customer generator to either avoid raising the error
or to provide an implementation.