Adapt Revision#each_modified_object to fix writing PDFs with >=2 revisions
When

* a document contains two or more revisions and
* those revisions originally use cross-reference tables and
* the document is written non-incrementally with the `optimize: true`
  option that adds a cross-reference table,

the Writer moves all added cross-reference dictionary objects into the
last revision. This leads to a file containing a mixture of
cross-reference tables and streams which doesn't work with Adobe Reader
and other viewers.

We fix this by making an exception for and ignoring the structural
dictionary objects of object and cross-reference streams in
Revision#each_modified_object. However, when writing incrementally we
need to move all modified or added/deleted objects into the appended
revision, so we add a new `all` argument to distinguish these two cases.