Support AFNumber_Format Javascript action for single line text fields
There are a variety of Javascript methods to format field values on an
interactive form after it has changed, e.g. for adding a currency
symbol.

Since HexaPDF doesn't (and won't) include a full Javascript engine, it
is problematic when filling out a form that uses those Javascript
methods.

This change implements one of the Javascript methods, AFNumber_Format.
This is done in a very simplistic matter: The Javascript string has to
strictly follow a certain formatting to be usable. A regular expression
is used to extract all the arguments and then process them accordingly.