Add Layout::PageStyle and update composer to use it
The new PageStyle class encapsulates all information about a page that
is needed by the composer.

The composer itself now uses page styles underneath although the old and
easy interface for quick usage still exists.

To make it possible to define a custom page style for the first page,
there is now an option for skipping page creation.