Handle potentially recoverable Zlib errors
It seems that some flate encoded streams that are invalid can be handled
more gracefully by providing as much of the original output as possible.
The result may still be invalid but not more invalid than before.

In some situations this may yield the complete and correct output if
only the end of the encoded stream is garbled or missing.

So this change should only improve the situation. If this is not wanted,
a configuration option can be used to restore the old behaviour.