Add Document::Layout#method_missing for convenient box creation
The ChildrenCollector already uses method_missing for conveniently
creating any kind of box, including ones registered through the
configuration.

By adding this functionality to the Layout class itself, that class also
gains this convenience and it makes it easy to further spread the
convenience methods to the Composer class in another step.