Fix AES decryption for empty strings and streams
When an empty string is encrypted with AES in CBC mode, the result
should be a 32 byte string, containing the initialization vector of 16
bytes and the padding of the 16 bytes.

There are, however, invalid files which will just have empty
strings/streams, without the IV or padding. This change makes HexaPDF
just return an empty string in such cases.