Fix Layout::InlineBox to work for all kinds of boxes
The InlineBox class is rather simple right now and doesn't work at all
if more demanding boxes are wrapped, like a TextBox.

To fully support all kinds of boxes, the wrapped box needs to be fit
inside a frame so that the correct width/height can be determined. This
change does exactly this.