Fix Layout::ListBox to take different final position into account
The list box fits its content into custom frames based on the position
in the provided frame. However, the final position of the list box may
be different, e.g. because it was absolutely positioned or due to being
centered.

This is currently not taken into account, the drawing always happens as
if in the initial position. By correctly translating the drawing
operations this is fixed.