Add Importer.copy for completely copying a PDF object
The Object#deep_copy method only copies direct values while leaving
indirect objects as they are. This is usually what one wants.

However, sometimes it might be useful to copy a PDF object *and* all
referenced indirect objects (e.g. when copying a page that gets modified
somehow). And this is what this new method is for.