Add Layout::Style::Border#draw_on_bounds
Usually, the border is drawn inside the given bounds. This corresponds
to HexaPDF's box model where everything drawn for a box is drawn inside
its bounds.

However, borders of table cells are much easier to handle if they are
drawn on the cell bounds since then they overlap and it is easier to
reason what happens. So this functionality is primarily implemented for
use by table cells.