Fix Document#wrap to correctly handle sub-type mapping when type is missing
It type and sub-type are known, the sub-type mapping based on type and
sub-type is working correctly.

In case type is missing, the sub-type mapping requires that all required
keys are set. This currently works if the type itself is not required
but fails otherwise. Case in point: HexaPDF::Type::FontTrueType which
has a mapping ':Font->:TrueType' as well as 'nil->:TrueType'. However,
the latter will never work since :Type is required.

To fix this the check for the required keys will now omit the :Type key.