Take width of single line text field widget into account when auto-sizing
The current implementation only takes care of making the text not too
tall for the widget. However, if the field value is longer it is
truncated by HexaPDF whereas other PDF libraries/apps shrink the font
size.

This change adjusts HexaPDF's behaviour to be more in line with other
common PDF libraries/apps.