Fix Page#flatten_annotations to correctly handle scaled appearances
While appearances are already scaled to fit into the annotation
rectangle, they are scaled from the origin. This works if the
appearance's bounding box has its bottom-left corner coinciding with the
origin but fails otherwise.

This can be fixed by scaling from the bottom-left corner of the bounding
box instead of from the origin.