Change Layout::Frame.new to accept a context argument
Some box classes may need to access the context in which they are laid
out during fitting. For example, a box may need to create other boxes.

The required context object is either a page or Form XObject instance.
Through it the HexaPDF document can be accessed.

There might be other similar changes in the future, for example, if it
is found necessary to also know box hierarchy.