Don't set dictionary default values if the type class might be wrong
It sometimes happens that an object of one type is wrapped by another
type class. For example, a page object might be wrapped by the page tree
node class.

When wrapping the object, the default values for required entries are
set. This can lead, in certain situations with certain viewers, to
problems. For example, if a page object is wrapped by the page tree node
class, the /Kids and /Count entries are set on the page object. That
shouldn't matter since /Type is still /Page. Howevever, some viewers
seem to consider such an object a page tree object instead of a page
object.

Additionally, such wrong wrappings shouldn't happen normally. However,
some files might contain additional objects with wrong references. For
example, if a file contains an additional catalog object and its /Pages
entry refers to an existing page in the real catalog object, that
reference might be resolved when optimizing the file, leading to the
problem.

This commit takes the precaution that the default values for the
required fields should only be set if the class' type matches the /Type
value, if it is set.