Add Style property mask_mode for control over to-be-removed region
The change allows one to fine-control the mask that defines the region
that gets removed after placing a box in a frame.

For example, it is now possible to set mask_mode=:fill to mask out the
whole of the current region. And if set to :none it is possible to
overlay boxes (think one big background image over which the content
boxes get laid out).