Change Glyph objects to take the font wrapper instead of the font
The three Glyph classes (Type1 glyphs, TrueType glyphs and the invalid
glyphs) currently take the font as first argument on initialization.

With this change the font wrapper is expected as the first argument.
Using the font wrapper it is still possible to access the font while it
is also possible to access the HexaPDF document instance and through it
everything else when only given a glyph instance.