Fix Layout#formatted_text_box to handle the properties correctly
The properties for the box itself can be specified via the properties
argument. And the properties for an individual text part can be
specified via the :properties key. However, this value was assigned to
the properties variable, overriding the box properties and thus using a
false value for the box properties.