Update Object#validate to catch exceptions and convert them into messages
The #perform_validation methods continue if a problem doesn't have
consequences for later checks. However, this is only true for the method
itself and not for possible callers.

For example, if the Resources type has an invalid value for the /ProcSet
entry, the call to the super perform_validation yields an appropriate
message and then proceeds. However, this might lead to an exception in
the Resource's #perform_validation method because of the wrong type.

Such situations may arise for any implemented PDF type but shouldn't
lead to an exception. Therefore we catch the exception and convert it
into an appropriate validation message.