Provide common fitting implementation in Layout::Box#fit that delegates to #fit_content
The Box#fit method is the only method out of the trio #fit/#split/#draw
that does not yet delegate to a *_content method. The methods
\#split/#draw provide a common implementation while the *_content method
should be used by subclasses to implement the specifics.

This change now introduces a #fit_content method that does the same for
the #fit method and the common case of fitting into the current
region (as opposed to flow positioning).