Fix Type::FontSimple#to_utf8 to work in case of errors during retrieving of the font's encoding
If the ToUnicode mapping cannot be retrieved or if there is no mapping,
nil is returned. However, a call to #encoding can result in an error
that breaks the ||-chain and therefore the configuration
'font.on_missing_unicode_mapping' is never invoked.

One instance where this happens is if the ToUnicode mapping exists but
doesn't have a mapping for a character code.

Fix this by swallowing the errors from #encoding.