Add Document::Metadata and Type::Metadata
The Document::Metadata class allows one to easily read metadata from the
information dictionary, change it, and then write it back to the
information dictionary as well as to the XMP metadata stream.

While writing a limited form of XMP metadata is supported, reading from
the XMP metadata stream is not supported.