Fix the table drawing implementation so that debug drawing works correctly
The current implementation of the drawing code for table cells modifies
Frame::FitResult#x and #y to set correct location. However, this only
modifies the position (x,y) and does nothing for the mask.

The correct way to do this is to use the dx and dy arguments for
FitResult#draw.