Use the style's font for list box markers if possible
Currently, the ZapfDingbats font is used for drawing the circle and
square markers. This is a problem in the context of PDF/A files.

Therefore the list box style's font is now tried first and only if it
doesn't contain the necessary glyph, the ZapfDingbats font is used as
fallback.