Change font fall back to first use :base style's font and only then Times
Currently, the style retrieval routine of HexaPDF::Document::Layout
unconditionally falls back to the Times font if no font is set on a
style object. This is problematic for PDF/A files.

This changes the logic to use the font set on the :base style first and
only if no font has been set, Times is used.