Add hierarchical box information to the document layout engine
While it is possible to have container boxes, the boxes themselves have
no way of knowing their parent boxes. However, this information is
useful for a variety of purposes (e.g. debugging, and, in the future,
tagging).

This change adds the information about the parent boxes to the Frame
instances. Each box that will be laid out in a frame now has to be a
child box of the immediate parent box associated with the frame.

There is also a new helper method for creating child frames so as to
make it easy to pass on all the needed information to those frames. This
method is used by all box classes so that the hierarchical box
information is correctly created.