Fix ListBox to create independent marker boxes for all markers
The ListBox implementation currently creates a single marker box for all
markers if they are the same. However, this is problematic behaviour
because boxes should only be re-used after they are fit AND drawn.

The implementation is changed so that only the text fragment's are
cached which is no problem because they can't be modified.