Make Font::CMap::Parser lazily parse bfranges
The current implementation of the CMap parser uses a greedy approach
when parsing the bfrange entries. It walks through each range and adds
all code-to-Unicode mappings to the listing.

This behaviour has two problems:

* If the CMap contains many bfrange entries covering much of Unicode,
  many entries will be added but only a small part will actually be
  used.

* Some CMap define mappings to invalid UTF-16BE ranges in the surrogate
  pair range, for example:

  <D800> <D8FF> <D800>

  This usually happens when the CMap tries to map all available codes.
  In the above case, however, converting the given Unicode value to the
  string "\xD8\x00" with UTF-16BE encoding yields a string with an
  invalid encoding as the second byte pair for the surrogate is missing.
  And a HexaPDF error.

To avoid both problems the bfrange entries of a CMap are now lazily
evaluated. Only in the case of a query for the Unicode value for a
specific character code is the bfrange value converted to a UTF-16BE
value.